del(FinalFrame): FinalFrame 제거 및 Player 쪽 사용 코드 수정
- FinalFrame Test 부분 FrameTest 쪽으로 이동
- NormalFrame 이름 변경
- Frame 인터페이스 제거
